// iot-service/iot-angular/src/sdk/core/models/update-owner-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface UpdateOwnerDto {

  /**
   * Contact person name
   */
  contactPerson?: string;

  /**
   * Forwarding settings (JSONB)
   */
  forwardingSettings?: {
};

  /**
   * Industry type
   */
  industry?: string;

  /**
   * Owner name
   */
  name?: string;

  /**
   * SLA level
   */
  slaLevel?: 'Bronze' | 'Silver' | 'Gold' | 'Platinum';
}